Move CommandResult to a new package
