+++ src/main/java/seedu/address/logic/commands/find/FindInternshipCommand.java
@@ -3,7 +3,7 @@ package seedu.address.logic.commands.find;
 import static java.util.Objects.requireNonNull;
 
 import seedu.address.commons.core.Messages;
-import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.results.CommandResult;
 import seedu.address.model.Model;
 import seedu.address.model.item.field.NameContainsKeywordsPredicate;
