+++ src/main/java/seedu/address/logic/commands/list/ListSkillCommand.java
@@ -3,7 +3,7 @@ package seedu.address.logic.commands.list;
 import static java.util.Objects.requireNonNull;
 import static seedu.address.model.Model.PREDICATE_SHOW_ALL_ITEMS;
 
-import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.results.CommandResult;
 import seedu.address.model.Model;
 
 /**
