+++ src/main/java/seedu/address/logic/commands/results/CommandResult.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.results;
 
 import static java.util.Objects.requireNonNull;
