+++ src/main/java/seedu/address/logic/commands/stubs/ViewResumeCommandStub.java
@@ -1,7 +1,7 @@
 package seedu.address.logic.commands.stubs;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.results.CommandResult;
 import seedu.address.logic.commands.view.ViewCommand;
 import seedu.address.model.Model;
