+++ src/test/java/seedu/address/logic/commands/add/AddProjectCommandTest.java
@@ -11,7 +11,7 @@ import java.util.Arrays;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.results.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 
 import seedu.address.model.item.Project;
