+++ src/test/java/seedu/address/logic/commands/delete/DeleteInternshipCommandTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.results.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.item.Internship;
 import seedu.address.testutil.InternshipBuilder;
